split tournament fight command flow #4846
